Simplify sizing by requiring all blocks to be less than usize (either 2^32 on 32 bit machines, or 2^64 on 64 bit machines).
Honestly, we're dealing with u16 addresses. What's the point? :-)

+++ crates/resources/src/file/data.rs
@@ -102,7 +102,7 @@ impl TryFrom<RawContents> for Contents {
             }
         };
         let decompressed_data = decompressed_data.with_check(move |block| {
-            if block.size() != raw_contents.unpacked_size as u64 {
+            if block.size() != raw_contents.unpacked_size as usize {
                 return Err(io::Error::other("Decompressed data size mismatch").into());
             }
             Ok(())
